refactor: renames components and adjusts style and layouts
