+++ frontend/src/lib/components/content/projects/ProjectHeader.svelte
@@ -1,220 +1,124 @@
 <script lang="ts">
+	import type { Snippet } from 'svelte';
+	import placeholderImage from '$lib/assets/placeholder-image.jpg';
 	import Subheader from '../../primatives/Subheader.svelte';
-	import TagList from '../../primatives/TagList.svelte';
-	import List from '../../primatives/List.svelte';
 
-	// Interface for props
-	interface ProjectHeaderProps {
-		children?: any;
-		impact?: {
-			title: string;
-			content: string;
-		};
-		categories?: {
-			title: string;
-			tags: string[];
-		};
-		teamMembers?: {
-			title: string;
-			members: { name: string; role: string }[];
-		};
+	// Interface for props following ComponentRules.md patterns
+	interface ProjectBodyProps {
+		leadingContent?: Snippet;
+		trailingContent?: Snippet;
+		content?: string;
+		imageSrc?: string;
+		imageAlt?: string;
 	}
 
-	let { children, impact, categories, teamMembers }: ProjectHeaderProps = $props();
+	let { leadingContent, trailingContent, content, imageSrc, imageAlt }: ProjectBodyProps = $props();
 
 	// Dummy placeholder data for debugging
-	const defaultImpact = {
-		title: 'Impact',
-		content: '10X revenue growth to $300+ MRR'
-	};
+	const defaultContent =
+		'A short description of my role in the project, helping to set the context for the project body content. What did I do, report to and hpw was it received?';
+	const defaultStandfirst = 'A green field design system and multi-platform design system';
 
-	const defaultCategories = {
-		title: 'Categories',
-		tags: ['UX Design', 'Product Strategy', 'User Research', 'Prototyping', 'Design System']
-	};
-
-	const defaultTeamMembers = {
-		title: 'Team Members',
-		members: [
-			{ name: 'Sarah Chen', role: 'Lead Designer' },
-			{ name: 'Marcus Rodriguez', role: 'Frontend Developer' }
-		]
-	};
-
-	// Use provided props or fallback to defaults for debugging
-	const impactData = impact || defaultImpact;
-	const categoriesData = categories || defaultCategories;
-	const teamMembersData = teamMembers || defaultTeamMembers;
+	// Use provided props or fallback to defaults
+	const displayContent = content || defaultContent;
+	const displayImageSrc = imageSrc || placeholderImage;
+	const displayImageAlt = imageAlt || 'Project placeholder image';
 </script>
 
-<!-- header Impact snippet -->
-{#snippet headerImpact()}
-	<div class="header-card-root impact">
-		<Subheader text={impactData.title} color="inverse" />
-		<p class="impact-text">{impactData.content}</p>
+<!-- leadingContent snippet -->
+{#snippet defaultLeadingContent()}
+	<div class="project-body__leading">
+		<img src={displayImageSrc} alt={displayImageAlt} class="project-body__image" />
 	</div>
 {/snippet}
 
-<!-- header Categories snippet -->
-{#snippet headerCategories()}
-	<div class="header-card-root categories">
-		<Subheader text={categoriesData.title} />
-		<div class="categories-content">
-			<TagList tags={categoriesData.tags} />
+<!-- trailingContent snippet -->
+{#snippet defaultTrailingContent()}
+	<div class="project-body__trailing">
+		<p class="project-standfirst">{defaultStandfirst}</p>
+		<div class="project-body__content">
+			<Subheader text="Project Overview" />
+			<p>{displayContent}</p>
 		</div>
 	</div>
 {/snippet}
 
-<!-- header Team Members snippet -->
-{#snippet headerTeamMembers()}
-	<div class="header-card-root team">
-		<Subheader text={teamMembersData.title} />
-		<List items={teamMembersData.members} dual />
-	</div>
-{/snippet}
+<section class="project-body">
+	<!-- Leading Content -->
+	{#if leadingContent}
+		{@render leadingContent()}
+	{:else}
+		{@render defaultLeadingContent()}
+	{/if}
 
-<section class="project-header">
-	<!-- Header Impact -->
-	<div class="header-card-root impact">
-		<Subheader text={impactData.title} color="inverse" />
-		<p class="impact-text">{impactData.content}</p>
-	</div>
-
-	<!-- Header Categories -->
-	<div class="header-card-root categories">
-		<Subheader text={categoriesData.title} />
-		<div class="categories-content">
-			<TagList tags={categoriesData.tags} />
-		</div>
-	</div>
-
-	<!-- Header Team Members -->
-	<div class="header-card-root team">
-		<Subheader text={teamMembersData.title} />
-		<List items={teamMembersData.members} dual />
-	</div>
+	<!-- Trailing Content -->
+	{#if trailingContent}
+		{@render trailingContent()}
+	{:else}
+		{@render defaultTrailingContent()}
+	{/if}
 </section>
 
 <style>
-	.project-header {
+	.project-body {
 		display: grid;
 		grid-template-columns: 1fr;
-		grid-template-rows: auto auto auto;
 	}
 
-	/* Tablet: 767px - 1024px */
-	@media (min-width: 767px) {
-		.project-header {
-			grid-template-columns: 1fr 1fr;
-			grid-template-rows: auto auto;
-		}
-
-		.header-card-root.impact {
-			grid-column: 1;
-			grid-row: 1;
-		}
-
-		.header-card-root.categories {
-			grid-column: 2;
-			grid-row: 1;
-		}
-
-		.header-card-root.team {
-			grid-column: 1 / -1;
-			grid-row: 2;
-		}
+	.project-body__leading {
+		display: flex;
+		justify-content: center;
+		aspect-ratio: 16 / 9;
 	}
 
-	/* Desktop: 1025px+ */
-	@media (min-width: 1025px) {
-		.project-header {
-			grid-template-columns: 1fr 1fr 2fr;
-			grid-template-rows: auto;
-		}
-
-		.header-card-root.impact {
-			grid-column: auto;
-			grid-row: auto;
-		}
-
-		.header-card-root.categories {
-			grid-column: auto;
-			grid-row: auto;
-		}
-
-		.header-card-root.team {
-			grid-column: auto;
-			grid-row: auto;
-		}
-	}
-
-	.header-card-root {
-		background: rgb(var(--bg-surface));
-		border: 1px solid rgb(var(--border-primary));
+	.project-body__image {
+		width: 100%;
 		border-radius: var(--radius-md);
-		padding: var(--space-xl);
+		box-shadow: var(--shadow-sm);
+		object-fit: cover;
+	}
+
+	.project-body__trailing {
 		display: flex;
 		flex-direction: column;
-		gap: var(--space-md);
-		transition: box-shadow 0.2s ease;
-		color: var(--fg-text-primary);
+		justify-content: center;
+		gap: var(--space-lg);
+		padding: var(--space-xl) var(--space-0);
 	}
 
-	.header-card-root.impact {
-		background-color: var(--bg-primary);
-		color: var(--fg-text-inverse);
+	.project-body__content {
+		max-width: 70ch;
 	}
 
-	.impact-text {
-		font-size: var(--fs-large-clamped);
-		font-weight: var(--fw-bold);
-		line-height: var(--lh-snug);
-		max-width: var(--width-prose-sm);
-	}
-
-	.team-member {
-		display: flex;
-		flex-direction: column;
-		gap: var(--space-2xs);
-	}
-
-	.team-member__name {
+	.project-body__content p {
 		font-size: var(--fs-300);
-		font-weight: var(--fw-semibold);
 		color: var(--fg-text-primary);
+		font-weight: var(--fw-regular);
+		line-height: var(--lh-body-text);
+	}
+
+	.project-standfirst {
+		font-size: var(--fs-600);
+		color: var(--fg-text-primary);
+		font-weight: var(--fw-regular);
 		font-family: var(--font-family-main);
+		line-height: var(--lh-snug);
 	}
 
-	.team-member__role {
-		font-size: var(--fs-275);
-		color: var(--fg-text-secondary);
-		font-family: var(--font-family-main);
-		line-height: var(--lh-interface);
-	}
+	/* Desktop: 1024px+ */
+	@media (min-width: 1024px) {
+		.project-body {
+			grid-template-columns: 1fr 1fr 2fr;
+			gap: var(--space-0);
+		}
 
-	/* Debug styles - remove when done */
-	/* .project-header {
-		box-shadow: 0 0 0 4px solid red !important;
-		outline: 2px solid red !important;
-	}
+		.project-body__leading {
+			grid-column: 1 / 3;
+		}
 
-	.header-card-root {
-		box-shadow: 0 0 0 4px solid blue !important;
-		outline: 2px solid blue !important;
+		.project-body__trailing {
+			grid-column: 3;
+			padding: var(--space-xl) var(--space-xl);
+		}
 	}
-
-	.header-card-root.impact {
-		box-shadow: 0 0 0 4px solid green !important;
-		outline: 2px solid green !important;
-	}
-
-	.header-card-root.categories {
-		box-shadow: 0 0 0 4px solid orange !important;
-		outline: 2px solid orange !important;
-	}
-
-	.header-card-root.team {
-		box-shadow: 0 0 0 4px solid purple !important;
-		outline: 2px solid purple !important;
-	} */
 </style>
