+++ frontend/src/lib/components/primatives/List.svelte
@@ -1,25 +1,35 @@
 <script lang="ts">
 	interface ListProps {
 		items?: string[] | { name: string; role: string }[];
-		dual?: boolean;
+		itemFormat?: 'string' | 'structured';
 		type?: 'plain' | 'bullet' | 'number' | 'custom';
 		customStyle?: string;
 	}
 
-	let { items = [
-		"John Smith – Developer",
-		"Sarah Johnson – Designer", 
-		"Mike Chen – Product Manager",
-		"Emily Davis – QA Engineer",
-		"Alex Rodriguez – DevOps"
-	], dual = false, type = 'plain', customStyle }: ListProps = $props();
+	let {
+		items = [
+			'John Smith – Developer',
+			'Sarah Johnson – Designer',
+			'Mike Chen – Product Manager',
+			'Emily Davis – QA Engineer',
+			'Alex Rodriguez – DevOps'
+		],
+		itemFormat = 'string',
+		type = 'plain',
+		customStyle
+	}: ListProps = $props();
 </script>
 
-<ul class="list list--{type}" style={type === 'custom' && customStyle ? `list-style: ${customStyle}` : undefined}>
+<ul
+	class="list list--{type}"
+	style={type === 'custom' && customStyle ? `list-style: ${customStyle}` : undefined}
+>
 	{#each items as item}
-		<li>
-			{#if dual && typeof item === 'object'}
-				{item.name} – {item.role}
+		<li class="list__item list__item--{itemFormat}">
+			{#if itemFormat === 'structured' && typeof item === 'object'}
+				<span class="list__item-name">{item.name}</span>
+				<span class="list__item-separator">–</span>
+				<span class="list__item-role">{item.role}</span>
 			{:else}
 				{item}
 			{/if}
@@ -29,7 +39,6 @@
 
 <style>
 	.list {
-		margin: 0;
 		padding: 0;
 	}
 
@@ -49,5 +58,26 @@
 
 	.list--custom {
 		padding-left: 1.5em;
+	}
+
+	.list__item--structured {
+		display: flex;
+		align-items: baseline;
+		gap: 0.25em;
+	}
+
+	.list__item-name {
+		font-weight: 500;
+		white-space: nowrap;
+	}
+
+	.list__item-separator {
+		color: var(--fg-text-secondary, #666);
+		white-space: nowrap;
+	}
+
+	.list__item-role {
+		color: var(--fg-text-secondary, #666);
+		white-space: nowrap;
 	}
 </style>
